posts-iws/requirements.txt
annotated-types==0.7.0
blinker==1.8.2
click==8.1.7
python-dotenv==1.0.1
Flask==3.0.3
Flask-Cors==5.0.0
flask-log-request-id==0.10.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.2
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3


